+++ ProjektAPBDCw2/Projector.cs
@@ -5,7 +5,7 @@ public class Projector : Device
     public int Brightness { get; }
     public int MaxRange { get; }
     
-    public Projector(string name, Manufacturer manufacturer, int price, int brightness, int maxRange) : base(name, manufacturer, price)
+    public Projector(string name, Manufacturer manufacturer, float price, int brightness, int maxRange) : base(name, manufacturer, price)
     {
         Brightness = brightness;
         MaxRange = maxRange;
